Moved the ALIAS env to the entry point workflow.
