+++ prefect_docker/scripts/flows/dataset_workflows/cbs_ingestion.py
@@ -1,17 +1,16 @@
 import os
 
-from prefect import flow
+from prefect import flow, get_run_logger
 from prefect.orion.schemas.states import Completed, Failed
 from tasks.base_tasks import xml2json, dataverse_mapper, \
     dataverse_import, update_publication_date, add_workflow_versioning_url
 
 CBS_MAPPING_FILE_PATH = os.getenv('CBS_MAPPING_FILE_PATH')
 CBS_TEMPLATE_FILE_PATH = os.getenv('CBS_TEMPLATE_FILE_PATH')
-CBS_DATAVERSE_ALIAS = os.getenv('CBS_DATAVERSE_ALIAS')
 
 
 @flow
-def cbs_metadata_ingestion(file_path, version):
+def cbs_metadata_ingestion(file_path, alias, version):
     json_metadata = xml2json(file_path)
     if not json_metadata:
         return Failed(message='Unable to transform from xml to json.')
@@ -25,13 +24,15 @@ def cbs_metadata_ingestion(file_path, version):
     if not mapped_metadata:
         return Failed(message='Unable to store workflow version.')
 
-    import_response = dataverse_import(mapped_metadata, CBS_DATAVERSE_ALIAS)
+    fields = mapped_metadata['datasetVersion']['metadataBlocks']['citation'][
+        'fields']
+    split_path = file_path.split('/')[3]
+    doi = 'doi:10.57934/' + split_path.split('_')[0]
+    import_response = dataverse_import(mapped_metadata, alias,
+                                       doi)
     if not import_response:
         return Failed(message='Unable to import dataset into Dataverse.')
 
-    doi = import_response.json()['data']['persistentId']
-    fields = mapped_metadata['datasetVersion']['metadataBlocks']['citation'][
-        'fields']
     publication_date = next((field for field in fields if
                              field.get('typeName') == 'distributionDate'),
                             {})
