+++ prefect_docker/scripts/flows/entry_workflows/main_cbs_ingestion.py
@@ -1,13 +1,12 @@
 import os
-from prefect import flow, get_run_logger
-
+from prefect import flow
 import utils
-from flows.cbs_ingestion import cbs_metadata_ingestion
+from flows.dataset_workflows.cbs_ingestion import cbs_metadata_ingestion
 from flows.workflow_versioning.workflow_versioner import \
     create_ingestion_workflow_versioning
 
 CBS_METADATA_DIRECTORY = os.getenv('CBS_METADATA_DIRECTORY')
-
+CBS_DATAVERSE_ALIAS = os.getenv('CBS_DATAVERSE_ALIAS')
 
 @flow
 def cbs_ingestion_pipeline():
@@ -17,7 +16,7 @@ def cbs_ingestion_pipeline():
                                                    importer=True,
                                                    updater=True)
     utils.workflow_executor(cbs_metadata_ingestion, CBS_METADATA_DIRECTORY,
-                            version)
+                            version, CBS_DATAVERSE_ALIAS)
 
 
 if __name__ == "__main__":
