+++ prefect_docker/scripts/flows/entry_workflows/main_easy_ingestion.py
@@ -1,11 +1,12 @@
 import os
 import utils
 from prefect import flow
-from flows.easy_ingestion import easy_metadata_ingestion
+from flows.dataset_workflows.easy_ingestion import easy_metadata_ingestion
 from flows.workflow_versioning.workflow_versioner import \
     create_ingestion_workflow_versioning
 
 EASY_METADATA_DIRECTORY = os.getenv('EASY_METADATA_DIRECTORY')
+EASY_DATAVERSE_ALIAS = os.getenv('EASY_DATAVERSE_ALIAS')
 
 
 @flow
@@ -15,7 +16,7 @@ def easy_ingestion_pipeline():
                                                    importer=True,
                                                    updater=True)
     utils.workflow_executor(easy_metadata_ingestion, EASY_METADATA_DIRECTORY,
-                            version)
+                            version, EASY_DATAVERSE_ALIAS)
 
 
 if __name__ == "__main__":
