+++ prefect_docker/scripts/flows/entry_workflows/main_liss_ingestion.py
@@ -1,11 +1,12 @@
 import os
 import utils
-from prefect import flow, get_run_logger
-from flows.liss_ingestion import liss_metadata_ingestion
+from prefect import flow
+from flows.dataset_workflows.liss_ingestion import liss_metadata_ingestion
 from flows.workflow_versioning.workflow_versioner import \
     create_ingestion_workflow_versioning
 
 LISS_METADATA_DIRECTORY = os.getenv('LISS_METADATA_DIRECTORY')
+LISS_DATAVERSE_ALIAS = os.getenv('LISS_DATAVERSE_ALIAS')
 
 
 @flow
@@ -15,7 +16,7 @@ def liss_ingestion_pipeline():
                                                    importer=True,
                                                    updater=True)
     utils.workflow_executor(liss_metadata_ingestion, LISS_METADATA_DIRECTORY,
-                            version)
+                            version, LISS_DATAVERSE_ALIAS)
 
 
 if __name__ == "__main__":
